complete the test scenario
I added:

1) setup code which creates a student with 5 books and 3 gadgets
2) code which removes the books, adds them back, removes the gadgets, adds them back, and flushes.

At the end of this, there are twice as many books as we expect. The difference between this and the issue we're seeing at work is - at work, adding $entityManager->flush() between the book and gadget setup resolves the issue. With this code, I have no such luck.